Refactorizacion del codigo
