+++ src/facturacion/components/index.ts
@@ -1,0 +1,5 @@
+import FacturacionCedulaFilter from "./FacturacionCedulaFilter";
+import FacturacionFechaFilter from "./FacturacionFechaFilter";
+import FacturacionTable from "./FacturacionTable";
+
+export { FacturacionCedulaFilter, FacturacionFechaFilter, FacturacionTable };
