+++ src/shared/components/index.ts
@@ -1,0 +1,8 @@
+import Title from "./Title";
+import SubTitle from "./SubTitle";
+import EndSlot from "./slots/EndSlot";
+import CardSlot from "./slots/CardSlot";
+import Table from "./Table";
+
+export { Title, SubTitle, EndSlot, CardSlot, Table };
+
