+++ src/sucursales/components/index.ts
@@ -1,0 +1,4 @@
+import BranchFilters from "./BranchFilters";
+import BranchTable from "./BranchTable";
+
+export { BranchFilters, BranchTable };
