+++ src/sucursales/modals/index.ts
@@ -1,0 +1,5 @@
+import AddBranchModal from "./AddBranchModal";
+import EditModal from "./EditModal";
+import ConfirmModal from "./ConfirmModal";
+
+export { AddBranchModal, EditModal, ConfirmModal };
